move timeout logic to InstanceConnectionManager.connect
